+++ src/main/java/org/icgc/argo/workflow_management/controller/RunsController.java
@@ -1,10 +1,10 @@
 package org.icgc.argo.workflow_management.controller;
 
 import lombok.val;
-import org.icgc.argo.workflow_management.model.dto.RunsRequest;
-import org.icgc.argo.workflow_management.model.dto.RunsResponse;
-import org.icgc.argo.workflow_management.model.dto.WESRunConfig;
+import org.icgc.argo.workflow_management.controller.model.RunsRequest;
+import org.icgc.argo.workflow_management.controller.model.RunsResponse;
 import org.icgc.argo.workflow_management.service.WorkflowExecutionService;
+import org.icgc.argo.workflow_management.service.model.WESRunParams;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -23,16 +23,29 @@ public class RunsController {
   @Qualifier("nextflow")
   private WorkflowExecutionService nextflowService;
 
+  @Autowired
+  @Qualifier("wdl")
+  private WorkflowExecutionService wdlService;
+
   @PostMapping
   private Mono<RunsResponse> postRuns(@Valid @RequestBody RunsRequest runsRequest) {
+    val wesService = resolveWesType("nextflow");
 
     // create run config from request
     val runConfig =
-        WESRunConfig.builder()
+        WESRunParams.builder()
             .workflow_url(runsRequest.getWorkflowUrl())
             .workflow_params(runsRequest.getWorkflowParams())
             .build();
 
-    return nextflowService.run(runConfig);
+    return wesService.run(runConfig);
+  }
+
+  private WorkflowExecutionService resolveWesType(String workflowType) {
+    if (workflowType.equals("wdl")) {
+      return wdlService;
+    }
+
+    return nextflowService;
   }
 }
