+++ src/main/java/org/icgc/argo/workflow_management/service/WorkflowExecutionService.java
@@ -1,11 +1,11 @@
 package org.icgc.argo.workflow_management.service;
 
-import org.icgc.argo.workflow_management.model.dto.RunsResponse;
-import org.icgc.argo.workflow_management.model.dto.WESRunConfig;
+import org.icgc.argo.workflow_management.controller.model.RunsResponse;
+import org.icgc.argo.workflow_management.service.model.WESRunParams;
 import reactor.core.publisher.Mono;
 
 public interface WorkflowExecutionService {
-  Mono<RunsResponse> run(WESRunConfig params);
+  Mono<RunsResponse> run(WESRunParams params);
 
   Mono<String> cancel(String runId);
