+++ src/test/java/org/icgc/argo/workflow_management/WorkflowManagementApplicationTests.java
@@ -9,8 +9,6 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 public class WorkflowManagementApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
-
+  @Test
+  public void contextLoads() {}
 }
